requirements.txt
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
colorlog==6.7.0
Cython==0.29.36
dnspython==2.3.0
docopt==0.6.2
idna==3.4
lxml==4.9.3
mutagen==1.46.0
numpy==1.25.1
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pipreqs==0.4.13
pycryptodomex==3.18.0
pymongo==4.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
tqdm==4.65.0
tweepy==4.14.0
tzdata==2023.3
urllib3==2.0.3
vidgear==0.3.0
webcolors==1.13
websockets==11.0.3
yarg==0.1.9
yt-dlp==2023.7.6
pdf2image
beautifulsoup4